requirements.txt
streamlit
pyaudio
mss
Pillow
imagehash
openai
requests
volcengine